+++ src/model/cst.py
@@ -66,6 +66,7 @@ class CST(nn.Module):
 class CSTShared(nn.Module):
     def __init__(
         self,
+        image_encoder="incepion_v3",
         dim=512,
         finetune_image_encoder=False,
         max_transformations=12,
@@ -76,7 +77,7 @@ class CSTShared(nn.Module):
         super().__init__()
 
         self.image_encoder = ImageEncoder(
-            name="inception_v3",
+            name=image_encoder,
             finetune=finetune_image_encoder,
             output_dim=dim,
             batch_norm=False,